+++ esp_tikz/src/initial_solution_duration_pdf_plotter.cpp
@@ -67,7 +67,7 @@ std::shared_ptr<PgfAxis> InitialSolutionDurationPdfPlotter::createInitialSolutio
 
   for (const auto& name : config_->get<std::vector<std::string>>("experiment/planners")) {
     auto plot = createInitialSolutionDurationPdfPlot(name);
-    plot->options.fillOpacity = config_->get<double>("initialSolutionPlots/combinedFillOpacity");
+    plot->options.fillOpacity = config_->get<float>("initialSolutionPlots/combinedFillOpacity");
     plot->options.lineWidth = config_->get<double>("initialSolutionPlots/lineWidth");
     axis_->addPlot(plot);
   }
